# refactor: stream asset ids for library queue jobs