Queue asset ids to redis via `.stream()` instead of the old pagination implementation.