@etnoy 2025-03-06
Benchmark done. 

This PR takes 10 seconds, main takes 8 seconds to import 28858 assets. This is without post sync jobs as I disabled queuePostSyncJobs to focus only on the library service.

10 vs 8 seconds is a 20% difference but I don't know if that is within the margin of error here. The disk might be spun up/better cached on the second run.

PR:

immich_server            | [Nest] 3015  - 03/06/2025, 10:34:51 PM     LOG [Api:LibraryService~yktqgwl5] Starting to scan library 3a65bd75-b36b-4126-aba9-58baf90131d4
immich_server            | [Nest] 3015  - 03/06/2025, 10:34:51 PM   DEBUG [Api:LoggingInterceptor~yktqgwl5] POST /api/libraries/3a65bd75-b36b-4126-aba9-58baf90131d4/scan 204 18.86ms ::ffff:172.18.0.6
immich_server            | [Nest] 3007  - 03/06/2025, 10:34:51 PM   DEBUG [Microservices:LibraryService] Validating import paths for library 3a65bd75-b36b-4126-aba9-58baf90131d4...
immich_server            | [Nest] 3007  - 03/06/2025, 10:34:51 PM     LOG [Microservices:LibraryService] Starting disk crawl of 1 import path(s) for library 3a65bd75-b36b-4126-aba9-58baf90131d4...
immich_server            | [Nest] 3007  - 03/06/2025, 10:34:52 PM     LOG [Microservices:LibraryService] Library 3a65bd75-b36b-4126-aba9-58baf90131d4 is empty, no need to check assets
immich_server            | [Nest] 3007  - 03/06/2025, 10:34:54 PM     LOG [Microservices:LibraryService] Crawled 10000 file(s) so far: 10000 of current batch of 10000 will be imported to library 3a65bd75-b36b-4126-aba9-58baf90131d4...
immich_server            | [Nest] 3007  - 03/06/2025, 10:34:56 PM     LOG [Microservices:LibraryService] Crawled 20000 file(s) so far: 10000 of current batch of 10000 will be imported to library 3a65bd75-b36b-4126-aba9-58baf90131d4...
immich_server            | [Nest] 3007  - 03/06/2025, 10:34:57 PM     LOG [Microservices:LibraryService] Crawled 28858 file(s) so far: 8858 of current batch of 8858 will be imported to library 3a65bd75-b36b-4126-aba9-58baf90131d4...
immich_server            | [Nest] 3007  - 03/06/2025, 10:34:57 PM     LOG [Microservices:LibraryService] Finished disk crawl, 28858 file(s) found on disk and queued 28858 file(s) for import into 3a65bd75-b36b-4126-aba9-58baf90131d4
immich_server            | [Nest] 3007  - 03/06/2025, 10:35:00 PM     LOG [Microservices:LibraryService] Imported 10000 (10000 done so far) file(s) into library 3a65bd75-b36b-4126-aba9-58baf90131d4
immich_server            | [Nest] 3007  - 03/06/2025, 10:35:00 PM     LOG [Microservices:LibraryService] Imported 10000 (20000 done so far) file(s) into library 3a65bd75-b36b-4126-aba9-58baf90131d4
immich_server            | [Nest] 3007  - 03/06/2025, 10:35:01 PM     LOG [Microservices:LibraryService] Imported 8858 (28858 done so far) file(s) into library 3a65bd75-b36b-4126-aba9-58baf90131d4


28858 assets in 10 seconds

main:

immich_server            | [Nest] 3119  - 03/06/2025, 10:41:11 PM     LOG [Api:LibraryService~2cgj7g64] Starting to scan library 3154854e-1f2c-4135-be54-24004daff734
immich_server            | [Nest] 3119  - 03/06/2025, 10:41:11 PM   DEBUG [Api:LoggingInterceptor~2cgj7g64] POST /api/libraries/3154854e-1f2c-4135-be54-24004daff734/scan 204 17.38ms ::ffff:172.18.0.6
immich_server            | [Nest] 3111  - 03/06/2025, 10:41:11 PM   DEBUG [Microservices:LibraryService] Validating import paths for library 3154854e-1f2c-4135-be54-24004daff734...
immich_server            | [Nest] 3111  - 03/06/2025, 10:41:11 PM     LOG [Microservices:LibraryService] Starting disk crawl of 1 import path(s) for library 3154854e-1f2c-4135-be54-24004daff734...
immich_server            | [Nest] 3111  - 03/06/2025, 10:41:11 PM     LOG [Microservices:LibraryService] Library 3154854e-1f2c-4135-be54-24004daff734 is empty, no need to check assets
immich_server            | [Nest] 3111  - 03/06/2025, 10:41:14 PM     LOG [Microservices:LibraryService] Crawled 10000 file(s) so far: 10000 of current batch of 10000 will be imported to library 3154854e-1f2c-4135-be54-24004daff734...
immich_server            | [Nest] 3119  - 03/06/2025, 10:41:14 PM   DEBUG [Api:LoggingInterceptor~pfqibi8f] GET /api/jobs 200 16.13ms ::ffff:172.18.0.6
immich_server            | [Nest] 3111  - 03/06/2025, 10:41:14 PM     LOG [Microservices:LibraryService] Crawled 20000 file(s) so far: 10000 of current batch of 10000 will be imported to library 3154854e-1f2c-4135-be54-24004daff734...
immich_server            | [Nest] 3111  - 03/06/2025, 10:41:15 PM     LOG [Microservices:LibraryService] Crawled 28858 file(s) so far: 8858 of current batch of 8858 will be imported to library 3154854e-1f2c-4135-be54-24004daff734...
immich_server            | [Nest] 3111  - 03/06/2025, 10:41:15 PM     LOG [Microservices:LibraryService] Finished disk crawl, 28858 file(s) found on disk and queued 28858 file(s) for import into 3154854e-1f2c-4135-be54-24004daff734
immich_server            | [Nest] 3111  - 03/06/2025, 10:41:18 PM     LOG [Microservices:LibraryService] Imported 10000 (10000 done so far) file(s) into library 3154854e-1f2c-4135-be54-24004daff734
immich_server            | [Nest] 3111  - 03/06/2025, 10:41:18 PM     LOG [Microservices:LibraryService] Imported 8858 (28858 done so far) file(s) into library 3154854e-1f2c-4135-be54-24004daff734
immich_server            | [Nest] 3111  - 03/06/2025, 10:41:19 PM     LOG [Microservices:LibraryService] Imported 10000 (20000 done so far) file(s) into library 3154854e-1f2c-4135-be54-24004daff734

28858 assets in 8 seconds